+++ src/app/(pages)/projetos/espaco-icosaedrico/album/page.jsx
@@ -1,0 +1,37 @@
+"use client";
+import { useState, useEffect } from "react";
+import { fetchSlideData } from "@/data/mutations";
+import Image from "next/image";
+
+const endpoint = process.env.API_URL_DEV + "/galeria/";
+
+const Page = () => {
+  const [images, setImages] = useState([]);
+
+  useEffect(() => {
+    fetchSlideData(setImages);
+  }, []);
+
+  return (
+    <main className="pt-[10em]">
+      <div className="lg:grid grid-cols-2 gap-10 grid-flow-row">
+        {images.map((e, index) => {
+          return (
+            <div className="lg:m-2 m-4" key={index}>
+              <Image
+                src={endpoint + `${images[index].id}`}
+                width={1200}
+                height={1200}
+                className="w-full lg:h-[480px] h-[360px] object-cover"
+                alt={images[index].description}
+              />
+              <h2 className="text-2xl lg:py-2 pb-2">{images[index].description}</h2>
+            </div>
+          );
+        })}
+      </div>
+    </main>
+  );
+};
+
+export default Page;
