+++ src/pages/Default/index.jsx
@@ -1,21 +1,17 @@
 import React from "react";
-import Button from "../Button";
-import Link from "next/link";
 
-const Hero = ({ heading, message }) => {
+const DefaultPage = ({ heading, message, children }) => {
   return (
-    <div className="flex items-center justify-center h-screen mb-12 bg-fixed bg-center bg-cover custom-image">
+    <div className="flex items-center justify-center h-screen mb-12 bg-fixed bg-center bg-cover custom-image-2">
       {/* Overlay */}
       <div className="absolute top-0 left-0 right-0 bottom-0 bg-black/75 z-[2] h-screen" />
       <div className="p-5 text-white z-[2] mt-[-2.5rem]">
         <h2 className="text-5xl font-bold">{heading}</h2>
         <p className="text-xl py-5">{message}</p>
-        <Link href={"/sobre"}>
-          <Button message={"Descubra mais sobre mim"} />
-        </Link>
+        <div className="mx-auto max-w-[54em]">{children}</div>
       </div>
     </div>
   );
 };
 
-export default Hero;
+export default DefaultPage;
